requirements.txt
torch>=2.1
torchvision
diffusers>=0.30.0
transformers>=4.40.0
accelerate>=0.30.0
peft>=0.11.0
safetensors
opencv-python
pillow
tqdm
numpy
scipy